[CI] Always clean up release channels
Previously, we deleted release channels at the end of successful
release pipeline runs. However, failed and canceled runs would leave
channels behind. Over time, this can accumulate a lot of extra
channels, which can (with the current implementation of things) slow
down build operations.

Here, we remove this responsibility from the Buildkite pipeline, and
instead allow Expeditor to handle it. This allows us to remove the
offending channel all the time, regardless of what happens in the
pipeline.

Signed-off-by: Christopher Maier <cmaier@chef.io>