[PR Feedback] Clean up a few extra acceptance references
This was mainly in our e2e tests. Longer term, we'll want to invest in
creating a much clearer distinction between the Builder we're pulling
release artifacts from, and the Builder (if any) that we want to
interact with.

This is particularly important with tests that might upload things to
Builder. As such, we're explicitly leaving the bulkupload test pointed
to Acceptance.

Signed-off-by: Christopher Maier <cmaier@chef.io>