Remove deploy_mac.sh and deploy_mac_launcher.sh
These were overdue for removal, being a relic of our old Travis CI
pipeline. `deploy_mac.sh` used the `bintray-publish` package, though,
which is now gone. `deploy_mac_launcher.sh` was the only thing that
called `deploy_mac.sh`, and nothing called it, so we'll just ditch
that while we're at it.

Signed-off-by: Christopher Maier <cmaier@chef.io>