[CI] Sign files using `gpg --yes`
We've seen intermittent failures to upload `manifest.json` files
because when we go to sign it, `gpg` says there's already a
`manifest.json.asc` file present, and asks whether we want to
overwrite it or not. As it's a tiny bit tricky to get STDIN in a
Buildkite stage, the stage ends up timing out.

Here, we simply add the `--yes` option, to automatically answer such
questions in the affirmative.

I'm not exactly sure how a `manifest.json.asc` could be present in a
clean runner, though.

Signed-off-by: Christopher Maier <cmaier@chef.io>