Add PIDS_FROM_LAUNCHER feature
To roll out the "no pidfiles" feature, we will put it behind a feature
flag. Users may opt-in to the feature by running their Supervisor with
`HAB_FEAT_PIDS_FROM_LAUNCHER=1` in the environment.

HOWEVER, they also have to be using a Launcher that can support the
new interaction. Because older Launchers would also just drop unknown
messages on the floor with no response, we have to be careful about
ultimately enabling this feature.

Rather than passing feature flags around and just determining our
behavior from that, this commit introduces the `ServicePidSource`
enum. It has a `determine_source` constructor which takes into account
both the feature flags the user has enabled, as well as the
capabilities of the Launcher being used. The resulting value is then
passed around. We do this because the Launcher capability test can
take a while (a second, currently, if the Launcher is old); we'd
rather do this test once at Supervisor startup time, rather than
_every_ time we want to know what the PID of a service is.

Once older Launcher are phased out, and getting PIDs from Launchers is
the accepted standard, we can remove all this extra
infrastructure. It's only here to help smooth the transition.

Signed-off-by: Christopher Maier <cmaier@chef.io>