catch unsuccessful chocolatey commands
An unsuccessful execution of a bare executable (as opposed to a
PowerShell cmdlet) in a PowerShell script won't fail the script. Every
execution of a binary needs to check the error level of the previous
command. So, this adds non-zero error level checks that will throw a
relatively informative message and will fail-exit the script so that the
release pipeline correctly reflects the success/fail status of the
package publishing.

Signed-off-by: Robb Kidd <robb@thekidds.org>