hab: support multiple idents and build vars in `hab pkg env`
This allows the user to run commands such as

    hab pkg env core/make core/gcc

The printed environment will contain the merged environment for
certain well-known variables. For unknown variables, the first package
to include the variable wins.

The --build flag now allows the user to print the build-relevant
environment variables for the given packages.

The --append-existing flag appends any existing values of the relevant
environment variables from the processes environment to the printed
environment.

Together, these features allow us to create ad-hoc development
environment using Habitat packages, which is nice for when developing
software locally:

   eval "$(hab pkg env core/gcc core/glibc core/openssl --build --append-existing)"

This is especially nice when trying to go through the
edit-compile-debug loop using local tools. In such a case a full
habitat package build is often slow without special code inside your
plan.sh specifically for dev builds.

If we are interesting in pursuing this, I think that to make this most
useful the path forward might look like:

- Review and exactly define the semantics of how variables are merged,
  build vs runtime variables, and transitive vs non-transitive
  dependencies.

- Rewrite some of the code in hab-plan-build to use this command
  directly so that we have only 1 implementation of this code.

Signed-off-by: Steven Danna <steve@chef.io>