avoid ctl_gateway panic on peer_addr().expect()
This change introduces a filter_map allowing for removal of the
peer_addr().expect() which resulted in panics. We filter out
instances where there is no peer address on the socket. The peer
address can be non-existant when a client disregards the protocol
such as from a half opened connection from `netcat -z`.

Signed-off-by: Jeremy J. Miller <jm@chef.io>