Add habitat-testing/test-probe app
This is a small webapp intended to help test the Supervisor. Right
now, it exposes the rendering context it is provided over an HTTP
endpoint, allowing a view into the Supervisor's inner workings from
outside.

It also has a handful of lifecycle hook implementations, some of which
take a long time to run. This has been handy for verifying async service
behavior.

This service could be augmented in the future to do other things, as
well.

Signed-off-by: Christopher Maier <cmaier@chef.io>