Generalize package builds in verify pipeline
Previously, everything we wanted to build lived in the `components`
directory. With the addition of testing service packages, this is no
longer the case.

The previous `build_component.{sh,ps1}` files assumed that everything
that needed building lived in `components`. Now, this is generalized;
instead of passing the name of the component to build, we pass the
path to the thing being built, which accommodates testing packages as
well.

The names of the scripts were changed to `build_package.{sh,ps1}` to
reflect this.

Signed-off-by: Christopher Maier <cmaier@chef.io>