Add some additional logging around update elections
We have observed rolling updates in leader topology groups stalling
due to the inability to determine who the leader of the service group
is (we need to know who the "real" leader is in order to determine who
should *not* win the update election).

This PR adds some logging output to help illustrate when this happens,
which will hopefully make this situation easier to troubleshoot in the
future.

Signed-off-by: Christopher Maier <cmaier@chef.io>