Remove PIDS_FROM_LAUNCHER feature flag
Now, the default behavior will be to get service PIDs directly from
the Launcher, rather than relying on pidfiles. However, to support
older Launchers, we will still fall back to using pidfiles if we
detect that the Launcher does not support the newer feature.

You can think of this as essentially inverting the feature flag, and
using the Launcher's abilities rather than an environment variable to
trigger it.

Signed-off-by: Christopher Maier <cmaier@chef.io>