Increase timeouts for tests
We've seen sporadic failures due to services taking longer than the
given timeout to be recognized as "up". This is generally due to
differences in how long it can take to download and install a new
service package.

Additionally, the `test-probe` service has intentionally long `init`
and `post-stop` hooks, which add to the amount of time it takes to
come back up.

Signed-off-by: Christopher Maier <cmaier@chef.io>