Ensure that signal handlers are installed first
Here, we ensure that our signal handlers are installed before we do
anything else.

Before, when running the Supervisor using a logging configuration file
with a refresh rate in it, our underlying logging library will start
up a refresh monitoring thread. This happened *before* we installed
our signal handlers, meaning that this refresh thread didn't get them.

As a result, when we try to restart the Supervisor by sending it a
`SIGHUP`, this refresh thread would cause the Supervisor and all its
services to completely shut down.

Though not strictly necessary at this point (because the Launcher
doesn't yet use `log4rs`), I've moved the signal handler
initialization in the Launcher up to the first instruction, as well.

Thanks to @stevendanna for the debugging assistance!

Signed-off-by: Christopher Maier <cmaier@chef.io>