Don't check for updates immediately with at-once updates
With at-once updates, we would start a service and then immediately
check to see if a new one was available. In certain pathological
cases (namely, quickly crashing-and-restarting services), this could
result in an excess of updater threads and over-consumption of network
resources.

This does not completely solve that issue, but should help ameliorate
it.

Signed-off-by: Christopher Maier <cmaier@chef.io>