Enable updater threads to work with tokio::time
This modifies the self-updater and service-updater threads to run
proper single-threaded, timer-enabled Tokio runtimes, in order to
allow the use of `tokio::time::delay_for` calls, both explicitly
within the updater threads themselves, but also from the `reqwest`
library, which is used to make HTTP calls to Builder from within these
threads.

Previously, a Tokio timer was not in place on these threads, which
would cause all such `delay_for` calls to panic.

This entered with #7409, but was quickly caught by end-to-end tests in
our build pipeline.

Thanks to @davidMcneil for debugging assistance!

Signed-off-by: Christopher Maier <cmaier@chef.io>