[Docs] Add basic `hab pkg download` file docs
We didn't have anything describing the formats of the files that `hab
pkg download` can take as input. There isn't currently a really great
place in the documentation structure to put this, so for now, it'll go
in the "Pattern Library", pending some broader reorganizations.

Ideally, we'd have all this in a `man` page or more extensive `hab pkg
download --help` documentation, but for now, we'll at least have it
documented somewhere.

Signed-off-by: Christopher Maier <cmaier@chef.io>