[CI] Allow CI pipeline to create Github PRs, releases
Our "finish release" pipeline does several tasks that need to interact
with Github in an authenticated way. This PR adjusts the pipeline to
provide the needed credentials to do this.

The main issue was that `hub` (which we use to create releases and
pull requests) needs additional configuration. While `hub` can be used
with the `GITHUB_TOKEN` of an application (i.e., Expeditor), we
must also set `GITHUB_USER` to the Github organization name of that
application if we wish to actually interact with the Github API with
it. See
https://github.com/github/hub/issues/2264#issuecomment-567241335 for
details.

Additionally, `maybe_run` was removed from the PR-creating
steps. Since creating a PR is an easily reversible step, there's no
harm in running it "for real".

A few other changes were made to how the PR-creating scripts are
executed to bring them closer together in terms of implementation.

Signed-off-by: Christopher Maier <cmaier@chef.io>