sup: set PID file mode to 0644
Users currently expect to be able to read the PID file from their
runtime hooks. In 5734cce7b we started using atomic writer for the PID
file to avoid corruption, but this meant that the default mode for the
file became 0600.

Note: I've restricted this change to non-Windows systems. If anyone
knows of what our Window's users might expect, I'm happy to extend
this PR to include a change for Windows as well.

Fixes #7351

Signed-off-by: Steven Danna <steve@chef.io>