Allow dev environment setup script to work locally or in Vagrant
Rather than sourcing the install script locally, we'll just use it via
`curl` like we normally do.

On one hand, it was nice to use the script from the
repository. However, you need internet access to do the setup anyway,
and this is a simple way to make the scripts work the same whether you
run them on your workstation or in a Vagrant machine.

Fixes #7454

Signed-off-by: Christopher Maier <cmaier@chef.io>