Remove time crate from Butterfly
With the update of the `time` crate to the 0.2 line, the `SteadyTime`
we were using was deprecated in favor of `Instant`. In light of that,
and given the headache of converting from the 0.1 line to the 0.2
line, as well as the fact that we're juggling three different time
crates across the project overall, it seems to be a good time to
simplify things. Since there doesn't seem to be a compelling reason to
stay with the `time` crate in Butterfly, we'll simply use `std::time`
instead.

All the changes in this commit are basically around moving from
`time::SteadyTime` to `std::time::Instant`, and moving from
`time::Duration` to `std::time::Duration`, with the associated API
changes (e.g., using `std::time::Duration#from_millis` instead of
`time::Duration#milliseconds`)

Of particular note is the change from `i64` to `u64` types in the
`server::timing` module. `time::Duration` accepted `i64`, but negative
durations have no place in our domain model. Converting everything
over to `u64` is easier, as well as being more correct.

Signed-off-by: Christopher Maier <cmaier@chef.io>