Switch from time::Timespec to std::time::SystemTime for Supervisor
Previously, we were using a `Timespec` to track when a Supervisor (the
process-specific one, not "The Supervisor (TM)") changed state. Since
this is no longer present in more recent versions of the `time` crate,
and we'd like to bring some sanity to our time usages, it was a good
opportunity to refactor.

The key purposes the `Timespec` was serving are as follows:

It provided the basis for our HTTP gateway's "seconds since the UNIX
Epoch" measurement of when a service changed state. Whether that is
the best representation for this information is another question; it's
currently part of our public interface, so that's what we're using.
As a result, we *have* to use `SystemTime` if we're going to stay with
the standard library (which we'd like to do at the moment, for
simplicity). This refactoring helps make that intention a bit more
obvious by implementing a private `since_epoch` method.

Additionally, it provided a kind of "change marker" in the `manager`
module's `service_states` map.

The end result is removing a bit more dependence on an external
library, and a hopefully clarified internal interface.

Signed-off-by: Christopher Maier <cmaier@chef.io>