tar exporter: remove `hab` dependency
For some reason, we had a dependency on the `hab` crate, even though
it wasn't being used. With the switch to compile-time generated
protobufs in that crate, however, the tar exporter Habitat package
refused to build because it did not have a build-time dependency on
`core/protobuf`.

Rather than adding that, we'll just remove the `hab` crate dependency!

Signed-off-by: Christopher Maier <cmaier@chef.io>