docker exporter: Add protobuf build-time dependency
This shouldn't really be needed, but the Docker exporter relies on the
`hab` crate for a few functions. None of them actually need any
protobuf-related code, but in order to compile the crate, we have to
have access to the protobuf compiler.

It's a bit of a bummer, and is something we should revisit, but for
now, this works.

Signed-off-by: Christopher Maier <cmaier@chef.io>