Remove time crate from the Supervisor
This is mainly mechanical conversions from `time::SteadyTime` to
`std::time::Instant`, as well as `time::Duration` to
`std::time::Duration`.

The more interesting work comes when we need to modify the
`habitat_sup::manager::commands::ProcessStatus` struct, which is
basically a mirror of the
`habitat_sup::manager::service::supervisor::Supervisor` struct after
going through a serialization to and from JSON. There are Rube
Goldbergian reasons for this current state of affairs, but it's what
we have. We need to do a bit more calculation to create a
`std::time::Duration` that represents how long the process has been in
its current state.

To help facilitate this, the `elapsed` field of the `ProcessStatus`
protobuf was changed from an `int64` to a `uint64`, mirroring the
change from `time::Duration` to `std::time::Duration`. Normally,
changing the type of a protobuf field is a bad idea, but in this case,
it turns out to be OK:

https://developers.google.com/protocol-buffers/docs/proto#updating

Signed-off-by: Christopher Maier <cmaier@chef.io>