Always write out service PID files
Previously in #7214 and then #7341, the Launcher and Supervisor were
modified so that the Launcher itself became the source of truth for
the PIDs of supervised services, rather than the contents of a file on
disk.

While this improves overall Supervisor reliability, a blanket removal
the PID files altogether can break things like `reconfigure` hooks,
which need a way to access a service's PID in order to do things like
send the service signals.

This PR re-introduces PID files for services, but _only_ as a means to
communicate with hooks. The Launcher is still the ultimate source of
the PID itself.

Fixes #7508

Signed-off-by: Christopher Maier <cmaier@chef.io>