Move Permissions enum to habitat_core::fs
Moving this into the core library just makes more sense. As a result,
we can also modify `AtomicWriter` to use this abstraction directly.

Furthermore, it allows us to push all the platform-specific behavior
into the type, which simplifies a lot of code by no longer requiring
conditional compilation.

There may be future opportunities to consolidate _all_
permission-related operations around this type, but for now, the
`AtomicWriter` is the focus.

Signed-off-by: Christopher Maier <cmaier@chef.io>