[PR Feedback] `AtomicWriter::new_with_permissions` constructor
This removes the need to create mutable instances of `AtomicWriter`
for the sole purpose of changing the permissions it uses.

As soon as another optional feature is added to `AtomicWriter`, we
should look into creating a Builder for it. Since we only have one
such feature now, though, a simple constructor will do the job nicely.

Signed-off-by: Christopher Maier <cmaier@chef.io>