core: move from users to nix crate
The nix crate is a bit more robust than the users create with respect
to its error handling. This moves us directly to the nix create
without changing any of the APIs in core::users.

This likely resolves #7298 for many users as nix correctly handles
ERANGE errors from getgrnam_r. However, the nix still enforces a
maximum buffer size of 16KB, so there is still a chance for a failure
caused by a large group. We could avoid this by calling libc directly
or working with upstream to make the buffer size configurable.

Signed-off-by: Steven Danna <steve@chef.io>