[REFACTOR] Push Windows Docker Engine check earlier
When exporting a container on Windows, the Docker daemon must be in
Windows mode. Previously, this check was buried deep in the logic,
after we had already done a lot of work to pull packages down. Now, we
perform this check first, thereby short circuiting a whole bunch of
unnecessary work, and cleaning up a bunch of logic.

Here, I'm trying to be a bit more intentional in how platform-specific
logic is organized in this crate, in order to make it more clear what
the general interfaces are, as well as what the *exact* platform
differences are, without littering conditional compilation directives
everywhere.

Additionally, in an effort to keep related code together, as well as
to start experimenting with better error organization, the error for a
failed platform check on Windows is actually kept with the platform
checking code itself, and converted into the generic Error we use
across the crate.

Signed-off-by: Christopher Maier <cmaier@chef.io>