Factor out an Identified trait for DockerBuilder and DockerImage
Both `DockerBuilder` and `DockerImage` had repeated logic for creating
the tags they use when creating an image, as well as when pushing and
deleting one.

One way to cut out the repetition is to factor the logic into a new
`Identified` trait that both structs can implement. It's admittedly
not the greatest name, but I suspect that it will ultimately be a
short-lived trait, pending some more work in progress.

Signed-off-by: Christopher Maier <cmaier@chef.io>