[CI] Fix e2e test that broke after the core plans refresh
Our `hab pkg download` test downloads `core/gzip` and then makes
assertions that all that package's transitive dependencies are also
downloaded. Because we don't yet have a great way to programmatically
extract that information from the artifact itself, the test hard-codes
knowledge of the dependencies.

After the most recent core plans refresh, the `core/gzip` package
gained a `core/zlib` dependency. The test didn't know about that, so
it ended up failing.

Here is `core/gzip` [before][1] the refresh, and here it is
[after][2].

(While we *can* extract the `TDEPS` file from the hart file with some
shell scripting, this really needs to be part of `hab pkg info`; it
would be more worthwhile to devote effort there. Thus, this PR takes
the expedient route in order to get the pipeline flowing again.)

[1]:https://bldr.habitat.sh/#/pkgs/core/gzip/1.9/20190115013612
[2]:https://bldr.habitat.sh/#/pkgs/core/gzip/1.10/20200306002325

Signed-off-by: Christopher Maier <cmaier@chef.io>