Handle container export report in backwards compatible way
Previously, we wrote out details of the exported container in
`last_docker_export.env` in the results directory. Now, we write out
things to `last_container_export.env`, but *also* continue to write a
copy of that information to `last_docker_export.env`, in case users
have any automation that relies on that path.

Eventualy we can remove this extra file, but in the meantime we won't
break anybody.

Signed-off-by: Christopher Maier <cmaier@chef.io>