Introduce Engine abstraction
Here we consolidate all the operations where we shell out to `docker`
behind a single unified Engine abstraction. In addition to reducing a
lot of duplication and centralizing common logic, this will also
provide a foundation to easily plug in alternative container-building
tools in the future.

Additionally, since the core removal and push logic now resides in the
`Engine`, what remained of those commands in `ContainerImage`
implementation were largely just concerned with UI and presentation
issues. As such, it's hard to see how they should belong to the
`ContainerImage` struct anymore (which seems fundamentally to be a
pure value struct). As such these have been moved up to the top level
of the crate with the rest of the main program flow.

(`ContainerImage::create_docker_config_file` was also moved for
similar reasons, though the use of `workdir` as both the the build
context and the location of the config file used for pushing images
makes for some still-awkward data sharing and access. This will be
further addressed in the future.)

Signed-off-by: Christopher Maier <cmaier@chef.io>