Remove ImageBuilder
As these refactorings have progressed, `ImageBuilder` has been
revealed to be the equivalent of Tom Smykowski in the movie Office
Space.

Now, the `Naming` struct generates the "expanded identifiers" that
`ImageBuilder` was via the `Identified` trait. This makes more sense,
in that all the identifier logic is consolidated in the `Naming`
struct.

A future commit will remove the `Identified` trait altogether, thus
fulfilling the prophecy foretold in 70a8acad.

Signed-off-by: Christopher Maier <cmaier@chef.io>