[PR Feedback] Add ok_log! macro and log nix errors
The `ok_log!` macro encapsulates the "lognore" idea that has been
floating around. It's basically for when you'd like to practically
ignore the `Err` variant of a `Result`, but still log the error for
visibility purposes.

Since some of these low-level Linux API calls can be kind of hairy,
it's useful to capture the errors somehow, even though we don't
necessarily want to plumb them through the entire API.

Signed-off-by: Christopher Maier <cmaier@chef.io>