Save spec in Service
Rather than storing every individual component of a `ServiceSpec`
inside a `Service` instance, we'll simply store a copy of the whole
spec itself.

Not only will this help clean up and streamline the `Service` struct,
it should make it easier to update service configuration on-the-fly in
the future by swapping in a new `ServiceSpec`.

Signed-off-by: Christopher Maier <cmaier@chef.io>