Allow dynamic changes to updater-related config for services
Previously, if you wanted to change the channel your service was
pulling from for its updates, or what builder instance it was pointed
to, or even whether it was updating or not, you would have to endure a
complete service restart. Technically, this should not be the case,
since service updates are managed completely outside of the running
service process itself; it is something the Supervisor does _around_
the service, not _to_ the service.

Now, changes to the `bldr_url`, `channel`, `update_strategy`, or
`update_condition` can be made to a service without incurring this
restart penalty. Changing other service parameters will still result
in a restart of a running service, but more of these may be made
dynamic in the future.

To take advantage of this now, you will need to use `hab svc load
--force`. However, be aware that this is not a "patchy" operation. If
you only wish to change the channel, you must supply all other
non-default service options in your load command; otherwise, default
values will be used, which will most likely result in a service
restart (as well as a potentially different configuration). We are
looking into providing a dedicated command for modifying the
configuration of running services that *will* be a patchy operation.

Signed-off-by: Christopher Maier <cmaier@chef.io>