Fix toToml serialization using `preserve_order` feature
Handlebars information is represented in terms of JSON objects. When
serializing to TOML, however, we can run into issues involving the
ordering of keys. With TOML, values must be serialized *before*
tables.

We can get around this by compiling `serde_json` with the
`preserve_order` feature, which maintains insertion order of keys in
JSON objects.

Signed-off-by: Christopher Maier <cmaier@chef.io>