hab-plan-build: don't fail on empty hooks dir
As of 2901d4d25fc0394793cf830b44c43c3aae5ec8ff, attempting to build a
package with an empty hooks directory would fail with the following:

    cp: can't stat '/src/habitat/hooks/*': No such file or directory

This is because when a shell glob does not match any files on the file
system, it is passed literally. Thus, cp is being passed the literal
string `/src/habitat/hooks/*`, which is not a file on the filesystem.

This is unfortunate because `hab plan init` creates an empty `hooks`
directory by default.

There are a few ways to fix this, including using find or setting the
nullglob option in Bash. I've opted to check for the existence of the
file again before doing anything with it.

Signed-off-by: Steven Danna <steve@chef.io>