ensure tdeps file is built with lower level deps at the bottom of the list
As the TDEPS metadata file is built, lowest level dependencies should be
at the bottom of the list because install hooks are rerun from bottom to
top. As long as each top level dep is not also included as a tdep, this
works today on windows (broken on linux). However if a top level dep is
also a tdep of a previously listed top level dep, the top level dep
remains at the top of the list. What needs to happen instead is for the
top level dep (that is also a tdep) to be removed from the top and then
appended to the bottom.

Linux adds an additional bug by sorting the TDEP (and all other
dependency metadata files) so all deps are listed in ident order
regardless of where they fall in the dependency tree. We do not perform
this sort on windows and from what I can tell, the sort has no value.

Signed-off-by: mwrock <matt@mattwrock.com>