Add `hab svc update` command
This new command can be used to update the Supervisor-specific
configuration for a running service. In contrast to `hab svc load
--force`, `hab svc update` is "patchy", in that you can specify only
the options you want to change, without having to also specify any
non-default values the service is currently configured with.

The service will not be restarted if the changes can be easily applied
to a running service.

Signed-off-by: Christopher Maier <cmaier@chef.io>