increase rumor file atomicity and durability
This change is a result of observations made while analyzing the state
of two independent Windows nodes which were panicking on attempting to
read the rumor file. The panic was occurring due to the 1st byte of the
file containing an invalid version number.

In one case I utilized the captured rumor file from the customer and
simply changed the first byte of the file from 0 to 2. The supervisor
was then able to read in the file and the content and carry on.

In the other case, I noticed that there was a .tmpxxxx file in the
supervisor's data directory alongside the rumor file and the content was
incompletely structured. This seems to be the vestige of the
AtomicWriter attempting to rename the temp file where the updates were
to be written to the live rumor file. The rename appeared to have
failed, potentially due to some external interrupter.

I therefore theorized that the header version `write` call could
have silently failed after reading this from rust docs:

"It is not considered an error if the entire buffer could not be written
to this writer.

An error of the ErrorKind::Interrupted kind is non-fatal and the write
operation should be retried if there is nothing else to do."

If the header version was not written and did not error, it would have
left the stream cursor at 0 byte position. The next call to `write`
would have been the header reserve which writes 64 zeros to the file.
That could have left the 1st byte of the file with a 0 as the version.

To help mitigate this theoretical scenario, I have replaced `write` with
`write_all`, which retries the write operation until the entire buffer
is written, even in the event that an EINTR occurs.

`write_all` implementation:
https://github.com/rust-lang/rust/blob/1d919c9377f4602d991ca1c7ba852e7555943740/src/libstd/io/mod.rs#L1399-L1411

In the other scenario, where a temp file existed, I realized that the
underlying implementation of `fs::rename` on Windows utilizes
`MoveFileExW` with only the `MOVEFILE_REPLACE_EXISTING` flag. This flag
allows for file overwrites but does not guarantee durability of the
rename operation to disk. The `MOVEFILE_WRITE_THROUGH` flag does. I
chose to call `MoveFileExW` directly on Windows, but also pass the
`MOVEFILE_WRITE_THROUGH` flag.

Underlying `fs::rename` implementation for Windows:
https://github.com/rust-lang/rust/blob/51eeabf505804fb3f2621626f5c39973640c514c/src/libstd/sys/windows/fs.rs#L712-L717

Docs for `MoveFileExW` and flags:
https://docs.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-movefileexw

Signed-off-by: Jeremy J. Miller <jm@chef.io>