[REFACTOR] Rework how changed service files are reported
First, we break out a `service_files` method, that returns an iterator
of all the group's current service files. This will be important in
coming commits, because before now, we did not have a way to access
this information.

Then, we turn `changed_service_files` into a `HashSet` and use that to
filter the list of all files.

Signed-off-by: Christopher Maier <cmaier@chef.io>