Write all service files when a service is loaded
Previously, because of how service files are tracked, you could start
up a service that did not have access to all its relevant service
files.

For instance, assume Supervisor A and Supervisor B are peered, and
that Supervisor A is running Service X, but Supervisor B is not. A
user uploads a file for Service X using `hab file upload`. The file is
written out to disk on Supervisor A, but not on Supervisor B, as
expected.

Then, the user loads Service X onto Supervisor B. The expectation
would be that the file is also written to disk there, but it is
not. Files are only written out at the time the file rumor is
received, and only if it is a "new" rumor. Since Supervisor B wasn't
running Service X when the file was uploaded, there was no managed
service present to cause the file to be written.

Uploading the file again with a new incarnation number would result in
Supervisor B writing out the file.

Here, we fix this by having the Supervisor write out any files for the
service that it knows about when the service is loaded, regardless of
whether or not they are "new" to the Supervisor.

Fixes #4877

Signed-off-by: Christopher Maier <cmaier@chef.io>