[CI] Bump macOS stage retry limits
Currently, our Anka backends seem to be oversubscribed globally,
resulting in timeouts to our stages that can cause pipelines to fail.

Out of an excessive paranoia, we're going to raise the retry limit to
a ridiculous number. We'll try and keep an eye on this behavior over
time so we can remove it when it's no longer required.

Signed-off-by: Christopher Maier <cmaier@chef.io>