[REFACTOR] Clarify users module interface
There are a large number of functions that are used in this module
that have platform-specific implementations (which themselves live in
separate submodules). The way we were structuring our `use` statements
obscured this fact, making it difficult to see the overall "shape" of
this module's interface. In particular, the fact that there are some
extra Unix-specific functions was not immediately obvious.

This commit restructures the `use` statements to make these points
explicit.

Signed-off-by: Christopher Maier <cmaier@chef.io>