[REFACTOR] Consolidate `setpgid` calls
We want to run all hooks in their own process groups for better
isolation. However, we had two separate implementations of this in the
codebase. Additionally, the one in the launcher didn't have any error
handling, while the one for the other hooks did.

Now, everything is consolidated into a single place, with a nicer
interface, and explanatory comments.

Signed-off-by: Christopher Maier <cmaier@chef.io>