Add `habitat-testing/supplemental-group-tester` package, with test
This new testing service is created to test whether or not the
Supervisor and Launcher both start hook files on Linux with the
service user's supplemental groups present. It does this by attempting
to read a file (set up external to Habitat, for our purposes) that the
`hab` user only has access to by a supplemental group membership.

Please read additional documentation in the plan and the accompanying
end-to-end test for more details.

Signed-off-by: Christopher Maier <cmaier@chef.io>