Don't fail key lookup if one particular file isn't readable
Previously, as we scanned all available keys looking for a specific
one, the entire process would halt if any one of the keys was
unreadable for any reason. Leaving aside how that might happen, it was
unfortunate, in that an unrelated unreadable key could prevent finding
the actual readable key you were looking for.

Here, we extract the file-checking logic to a separate function (with
unit tests) for an increase in clarity, while also handling I/O errors
by logging-and-continuing, rather than failing the entire search.

Fixes #7269

Signed-off-by: Christopher Maier <cmaier@chef.io>