[CI] Binlink `hab` earlier in the path for e2e tests
In our CI environment, `/usr/bin` comes ahead of `/bin` in the
path. We have Chef Workstation's `hab` binary linked to
`/usr/bin/hab`, which will mask anything we install and binlink to
`/bin/hab` by default.

Now, we link *our* `hab` into `/usr/bin`.

Additionally, we remove the `curlbash_hab` call when we set up our e2e
environment. That's only to get a `hab` on the box, but we'll already
have one because of the images and containers we're using in CI.

We also better document what the situation is in our
`test_install_script.bats` test.

Signed-off-by: Christopher Maier <cmaier@chef.io>