[REFACTOR] RingKey::get_secret_key_path -> cached_path
RingKeys are always secret, so there's no point in specifying
that. Additionally, the name helps to convey that we're looking for a
file that should already exist in the cache, rather than possibly just
generating the path at which we should look for that file.

Clean up some type parameters in related functions/methods as well.

Signed-off-by: Christopher Maier <cmaier@chef.io>