[REFACTOR] Extract a KeyCache struct for RingKey operations
`RingKey` (and all keys, really) are performing more than they need
to. In particular, keys are handling the logic to write themselves to
a cache, as well as find various revisions from the cache. This
suggests a formal `KeyCache` struct that can act as a "database" of
sorts for keys, leaving the key structs themselves to just deal with
things like encryption and decryption.

Here, we introduce `KeyCache` for `RingKey` operations. At the moment,
all the implementation remains in `RingKey`, but the public interface
is in terms of `KeyCache`. Further work is envisioned for pulling
other key operations into the `KeyCache`, allowing us to further
streamline and consolidate logic, and make it apparent what the actual
differences in behavior for each kind of key really are.

In the process `RingKey::get_latest_pair_for` was renamed to
`latest_cached_revision`; really, that should have been a separate
commit, but I got a little carried away :/

Signed-off-by: Christopher Maier <cmaier@chef.io>