[REFACTOR] Simplify write_file_from_str by removing tempfiles
We don't need to write the key to a file, then hash that file in order
to compare contents to a similar file already on disk. We can just
hash the string the key would generate.

This dramatically simplifies the logic around conditionally writing
out a key file, paving the way to extracting it to the `KeyCache` in
the future.

Also, stop returning a `PairType` from this function. `RingKey`s
are *always* secret, so this adds nothing.

Signed-off-by: Christopher Maier <cmaier@chef.io>