[REFACTOR] Remove RingKey::write_file_from_str
Instead, parse a `RingKey`, and then use the `KeyCache` to save it to
the filesystem.

Now, rather than having one complex function that does two things, we
have two simple functions that can be composed.

Also as part of this, change the setup of the `KeyCache` to be a bit
more principled.

Signed-off-by: Christopher Maier <cmaier@chef.io>