[REFACTOR] Remove KeyCache::ring_key_cached_path
This was only ever used in the `hab ring key export` command, and even
then, it was unnecessary. If the initial key lookup was successful,
then we already have all the information needed to print the key
contents to standard output; there was never a reason to make a second
call to access the file directly.

Signed-off-by: Christopher Maier <cmaier@chef.io>