[REFACTOR] Create FromStr implementations for keys using a macro
The logic from the `RingKey` `FromStr` implementation was essentially
the same as will be needed for all our keys, since they all share the
same basic structure. To facilitate this, we now have a
`from_str_impl_for_key` macro that does all the heavy lifting for us.

The main logic of the implementation is in the `key_bits` function,
which is essentially a streamlined version of the `RingKey`'s old
implementation. In order to make it generic in the way we need, a
rather silly `FromSlice` trait was introduced to expose the
`from_slice` method that our various `sodiumoxide` key types have, but
are not themselves exposed as a trait. To further save typing, the
`from_str_impl_for_key` macro generates those implementations as well,
since they're just implementation details that wouldn't otherwise have
to exist.

Signed-off-by: Christopher Maier <cmaier@chef.io>