[CI] Ensure that kernel2 packages get build appropriately
It seems that some recent changes in CI may have made it so that we
get `x86_64-linux` `hab` binaries running our builds for kernel2
packages, which, of course, results in not actually getting kernel2
packages.

I'm not yet clear on the mechanisms, but it seems we need to
explicitly use `${hab_binary}` rather than simply `hab` to run `hab
pkg path`, or else we get the wrong package back.

This PR does that, as well as cleans up logging, tightens up helper
function contracts around package targets, and adds a test that
ensures we are building the intended package target.

Signed-off-by: Christopher Maier <cmaier@chef.io>